+++ EnvironmentSimulator/code-examples/hello_world/hw4_external_ego_control.cpp
@@ -5,7 +5,7 @@ int main(int argc, char* argv[])
 {
     const char*            filename = argc > 1 ? argv[1] : "../resources/xosc/cut-in_external.xosc";
     SE_ScenarioObjectState state;
-    double                 z = 0.0f;
+    double                 z = 0.0;
 
     if (SE_Init(filename, 0, 1, 0, 0) != 0)
     {
@@ -14,19 +14,14 @@ int main(int argc, char* argv[])
 
     for (int i = 0; i < 500 && SE_GetQuitFlag() != 1; i++)
     {
-        SE_ReportObjectPos(SE_GetId(0), 8.0f, i, z, 1.57 + 0.01 * i, 0.0f, 0.0f);
+        SE_ReportObjectPos(SE_GetId(0), 8.0, i, z, 1.57 + 0.01 * i, 0.0, 0.0);
 
         SE_Step();
 
         // fetch position in terms of road coordinates
         SE_GetObjectState(SE_GetId(0), &state);
 
-        printf("road_id: %d s: %.3f lane_id %d lane_offset: %.3f z: %.2f\n",
-               state.roadId,
-               static_cast<double>(state.s),
-               state.laneId,
-               static_cast<double>(state.laneOffset),
-               static_cast<double>(state.z));
+        printf("road_id: %d s: %.3f lane_id %d lane_offset: %.3f z: %.2f\n", state.roadId, state.s, state.laneId, state.laneOffset, state.z);
 
         if (i == 100)
         {
