+++ EnvironmentSimulator/code-examples/ideal-sensors/ideal-sensors.cpp
@@ -13,10 +13,10 @@ int main(int argc, char* argv[])
     SE_Init("../resources/xosc/cut-in.xosc", 0, 1, 0, 0);
 
     int sensor_id = -1;
-    sensor_id     = SE_AddObjectSensor(0, 2.0f, 1.0f, 0.5f, 1.57f, 1.0f, 50.0f, 1.57f, MAX_HITS);
+    sensor_id     = SE_AddObjectSensor(0, 2.0, 1.0, 0.5, 1.57, 1.0, 50.0, 1.57, MAX_HITS);
     printf("Added sensor %d to object %s\n", sensor_id, SE_GetObjectName(SE_GetId(0)));
 
-    sensor_id = SE_AddObjectSensor(0, -1.0f, 0.0f, 0.5f, 3.14f, 0.5f, 20.0f, 1.57f, MAX_HITS);
+    sensor_id = SE_AddObjectSensor(0, -1.0, 0.0, 0.5, 3.14, 0.5, 20.0, 1.57, MAX_HITS);
     printf("Added sensor %d to object %s\n", sensor_id, SE_GetObjectName(SE_GetId(0)));
 
     for (int i = 0; i < 2000 && !(SE_GetQuitFlag() == 1); i++)
@@ -35,11 +35,7 @@ int main(int argc, char* argv[])
                 SE_ScenarioObjectState state;
 
                 SE_GetObjectState(objList[k], &state);
-                printf("object[%d] pos: (%.2f, %.2f) heading: %.2f\n",
-                       objList[k],
-                       static_cast<double>(state.x),
-                       static_cast<double>(state.y),
-                       static_cast<double>(state.h));
+                printf("object[%d] pos: (%.2f, %.2f) heading: %.2f\n", objList[k], state.x, state.y, state.h);
             }
         }
     }
